Add frontend-only authentication module: RegisterForm, LoginForm, LogoutButton, UserContext, and Navigator
