+++ index.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="zh">
+  <head>
+    <!--
+      index.html
+      本文件是项目的入口 HTML 文件，用于加载 React 应用。
+      它包含一个 id 为 "root" 的 div 元素作为 ReactDOM 的挂载点。
+    -->
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+    <title>图书馆管理系统</title>
+  </head>
+  <body>
+    <!-- React 应用挂载点 -->
+    <div id="root"></div>
+    <!-- Vite 模块入口: 加载 src/index.js -->
+    <script type="module" src="/src/index.jsx"></script>
+  </body>
+</html>
